Add Windows Forms support

+++ DirectX12GameEngine/HolographicGraphicsPresenter.cs
@@ -30,7 +30,7 @@ namespace DirectX12GameEngine
             {
                 case GameContextHolographic context:
                     CoreWindow coreWindow = context.Control;
-                    coreWindow.SizeChanged += (s, e) => SizeChanged?.Invoke(this, new SizeChangedEventArgs((int)e.Size.Width, (int)e.Size.Height));
+                    coreWindow.SizeChanged += (s, e) => SizeChanged?.Invoke(this, new SizeChangedEventArgs(e.Size.Width, e.Size.Height));
 
                     using (SharpDX.DXGI.Device dxgiDevice = GraphicsDevice.NativeDirect3D11Device.QueryInterface<SharpDX.DXGI.Device>())
                     {
